fix #396: hair switcher objects armature not disabled/hair can be summoned on top of the hair switched changing hair_list
